Move graph indices into own package
