+++ src/main/java/org/opentripplanner/routing/linking/SameEdgeAdjuster.java
@@ -11,6 +11,7 @@ import org.opentripplanner.routing.edgetype.TemporaryPartialStreetEdge;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
+import org.opentripplanner.routing.graph.index.Scope;
 import org.opentripplanner.routing.vertextype.StreetVertex;
 
 public class SameEdgeAdjuster {
